pyproject.toml
[project]
name = "Browser-Security-Bot"
version = "0.1.0"
description = "A discord bot for browser security research"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "discord-py>=2.5.0",
    "httpx>=0.28.1",
]
